[admin][sr] add user overview component
