+++ admin/app/components/user-detail-personal-information.js
@@ -1,13 +1,9 @@
 import Component from '@glimmer/component';
-import { tracked } from '@glimmer/tracking';
 // eslint-disable-next-line ember/no-computed-properties-in-native-classes
-import Object, { action, computed } from '@ember/object';
-import { validator, buildValidations } from 'ember-cp-validations';
-import { getOwner } from '@ember/application';
+import { action } from '@ember/object';
 import { inject as service } from '@ember/service';
+import { tracked } from '@glimmer/tracking';
 import get from 'lodash/get';
-
-import ENV from 'pix-admin/config/environment';
 
 const DISSOCIATE_SUCCESS_NOTIFICATION_MESSAGE = 'La dissociation a bien été effectuée.';
 
@@ -18,83 +14,7 @@ const typesLabel = {
   GAR: 'Médiacentre',
 };
 
-const Validations = buildValidations({
-  firstName: {
-    validators: [
-      validator('presence', {
-        presence: true,
-        ignoreBlank: true,
-        message: 'Le prénom ne peut pas être vide.',
-      }),
-      validator('length', {
-        min: 1,
-        max: 255,
-        message: 'La longueur du prénom ne doit pas excéder 255 caractères.',
-      }),
-    ],
-  },
-  lastName: {
-    validators: [
-      validator('presence', {
-        presence: true,
-        ignoreBlank: true,
-        message: 'Le nom ne peut pas être vide.',
-      }),
-      validator('length', {
-        min: 1,
-        max: 255,
-        message: 'La longueur du nom ne doit pas excéder 255 caractères.',
-      }),
-    ],
-  },
-  email: {
-    validators: [
-      validator('presence', {
-        presence: true,
-        ignoreBlank: true,
-        message: "L'e-mail ne peut pas être vide.",
-        disabled: computed.none('model.email'),
-      }),
-      validator('length', {
-        max: 255,
-        message: "La longueur de l'email ne doit pas excéder 255 caractères.",
-        disabled: computed.none('model.email'),
-      }),
-      validator('format', {
-        ignoreBlank: true,
-        type: 'email',
-        message: "L'e-mail n'a pas le bon format.",
-        disabled: computed.none('model.email'),
-      }),
-    ],
-  },
-  username: {
-    validators: [
-      validator('presence', {
-        presence: true,
-        ignoreBlank: true,
-        message: "L'identifiant ne peut pas être vide.",
-        disabled: computed.none('model.username'),
-      }),
-      validator('length', {
-        min: 1,
-        max: 255,
-        message: 'La longueur du nom ne doit pas excéder 255 caractères.',
-        disabled: computed.none('model.username'),
-      }),
-    ],
-  },
-});
-
-class Form extends Object.extend(Validations) {
-  @tracked firstName;
-  @tracked lastName;
-  @tracked email;
-  @tracked username;
-}
-
 export default class UserDetailPersonalInformationComponent extends Component {
-  @tracked isEditionMode = false;
   @tracked displayAnonymizeModal = false;
   @tracked displayDissociateModal = false;
   @tracked displayRemoveAuthenticationMethodModal = false;
@@ -105,29 +25,8 @@ export default class UserDetailPersonalInformationComponent extends Component {
 
   schoolingRegistrationToDissociate = null;
 
-  constructor() {
-    super(...arguments);
-    this.form = Form.create(getOwner(this).ownerInjection());
-  }
-
-  get canAdministratorAnonymizeUser() {
-    return this.args.user.email;
-  }
-
-  get externalURL() {
-    const urlDashboardPrefix = ENV.APP.USER_DASHBOARD_URL;
-    return urlDashboardPrefix && urlDashboardPrefix + this.args.user.id;
-  }
-
   get translatedType() {
     return typesLabel[this.authenticationMethodType];
-  }
-
-  @action
-  changeEditionMode(event) {
-    event.preventDefault();
-    this._initForm();
-    this.isEditionMode = !this.isEditionMode;
   }
 
   @action
@@ -152,37 +51,6 @@ export default class UserDetailPersonalInformationComponent extends Component {
     await this.args.user.save({ adapterOptions: { anonymizeUser: true } });
     await this.args.user.reload();
     this.toggleDisplayAnonymizeModal();
-  }
-
-  @action
-  cancelEdit() {
-    this._initForm();
-    this.isEditionMode = false;
-  }
-
-  @action
-  async updateUserDetails(event) {
-    event.preventDefault();
-
-    const { validations } = await this.form.validate();
-    if (!validations.isValid) {
-      return;
-    }
-    this.args.user.firstName = this.form.firstName.trim();
-    this.args.user.lastName = this.form.lastName.trim();
-    this.args.user.email = !this.form.email ? null : this.form.email.trim();
-    this.args.user.username = !this.form.username ? null : this.form.username.trim();
-
-    try {
-      await this.args.user.save();
-      this.notifications.success('L’utilisateur a été mis à jour avec succès.');
-      this.isEditionMode = false;
-    } catch (response) {
-      this.args.user.rollbackAttributes();
-      const messageValidationError =
-        response.errors[0].detail || "une erreur est survenue, vos modifications n'ont pas été enregistrées";
-      this.notifications.error(messageValidationError);
-    }
   }
 
   @action
@@ -215,12 +83,5 @@ export default class UserDetailPersonalInformationComponent extends Component {
       this.isLoading = false;
       this.toggleDisplayRemoveAuthenticationMethodModal(null);
     }
-  }
-
-  _initForm() {
-    this.form.firstName = this.args.user.firstName;
-    this.form.lastName = this.args.user.lastName;
-    this.form.email = this.args.user.email;
-    this.form.username = this.args.user.username;
   }
 }
